Added ConfigurationTask and EnergyConfTask These FireTasks server the purpose of modularizing the configuration workflow we currently have:
ConfigurationTask: Set up the configuration_dict and pass it to the fw_spec
EnergyConfTask: For each configuration in the configuration_dict, set up a Firework for first optimizing the geometry and then calculating the energy more precisely using a SCF calculation.

This commit is the initial draft of these FireTasks. Now proceeding with testing.