update: DimerNEBAnalysis Warn the user in case the from_dir method is unable to figure out the dimer based on the directory.
Note that the current way of doing this is rather dependent on the directory having a suitable name that contains the dimer indices between _'s. This is not too convenient. However, it would be more involved to figure out which indices correspond to the O-O dimer from the output files. Might want to think this through later and improve.