remove: naming scheme for get_wf_neb
The naming scheme for the NEB workflow still relied on reading the Cathode from a file on the cluster. This meant that it wasn't possible to submit the workflow from a jupyter notebook. Now the workflow has a rather generic default name, but the user can specify a more informative description via an input argument.